Re-enable compat build for libc and libutil
I don't remember why it was originally disabled, but now I can't think of
a good reason to deviate from unmodified sources.  The compat parts of
the libs are useless for rumprun, but due to .a semantics they'll just
not be linked into the final image, i.e. mostly harmless.